+++ app/Http/Controllers/Controller.php
@@ -27,6 +27,11 @@ class Controller extends BaseController
      */
     protected $include_private;
 
+    /**
+     * @var boolean Allow the entire collection to be returned ignoring pagination
+     */
+    protected $allow_entire_collection = false;
+
     public function __construct()
     {
         $this->hash = new Hash();
@@ -104,6 +109,7 @@ class Controller extends BaseController
             'sortable_config' => null,
             'searchable_config' => null,
             'enable_pagination' => true,
+            'allow_entire_collection' => false,
             'authentication_required' => false
         ],
         array $post = [
@@ -121,7 +127,14 @@ class Controller extends BaseController
         if ($get['parameters_config_string'] !== null) {
             foreach (
                 array_merge_recursive(
-                    ($get['enable_pagination'] === true ? Config::get('api.pagination.parameters') : []),
+                    ($get['enable_pagination'] === true ?
+                        (
+                            $get['allow_entire_collection'] === false ?
+                                Config::get('api.pagination.parameters') :
+                                Config::get('api.pagination.parameters-including-collection')
+                        ) :
+                        []
+                    ),
                     ($get['sortable_config'] !== null ? Config::get('api.sortable.parameters') : []),
                     ($get['searchable_config'] !== null ? Config::get('api.searchable.parameters') : []),
                     Config::get($get['parameters_config_string']),