+++ app/Http/Controllers/ItemCategoryController.php
@@ -126,6 +126,7 @@ class ItemCategoryController extends Controller
                 'sortable_config' => null,
                 'searchable_config' => null,
                 'enable_pagination' => false,
+                'allow_entire_collection' => $this->allow_entire_collection,
                 'authentication_required' => false
             ],
             [