+++ app/Http/Controllers/ResourceTypeController.php
@@ -26,6 +26,8 @@ use Illuminate\Support\Facades\Config;
  */
 class ResourceTypeController extends Controller
 {
+    protected $allow_entire_collection = true;
+
     /**
      * Return all the resource types
      *
@@ -46,7 +48,12 @@ class ResourceTypeController extends Controller
             Config::get('api.resource-type.sortable')
         );
 
-        $pagination = UtilityPagination::init(request()->path(), $total)->
+        $pagination = UtilityPagination::init(
+                request()->path(),
+                $total,
+                10,
+                $this->allow_entire_collection
+            )->
             setSearchParameters($search_parameters)->
             setSortParameters($sort_parameters)->
             paging();
@@ -146,6 +153,7 @@ class ResourceTypeController extends Controller
                 'sortable_config' => 'api.resource-type.sortable',
                 'searchable_config' => 'api.resource-type.searchable',
                 'enable_pagination' => true,
+                'allow_entire_collection' => $this->allow_entire_collection,
                 'authentication_required' => false
             ],
             [