+++ app/Http/Controllers/SummaryRequestController.php
@@ -57,6 +57,7 @@ class SummaryRequestController extends Controller
                 'sortable_config' => null,
                 'searchable_config' => null,
                 'enable_pagination' => false,
+                'allow_entire_collection' => $this->allow_entire_collection,
                 'authentication_required' => false
             ]
         );