+++ app/Http/Controllers/SummaryResourceTypeItemController.php
@@ -498,6 +498,7 @@ class SummaryResourceTypeItemController extends Controller
                 'sortable_config' => null,
                 'searchable_config' => 'api.resource-type-item.searchable',
                 'enable_pagination' => false,
+                'allow_entire_collection' => $this->allow_entire_collection,
                 'authentication_required' => false
             ]
         );