+++ resources/lang/en/pagination/parameters.php
@@ -7,5 +7,8 @@ return [
     'description-offset' => 'Start result for pagination',
 
     'title-limit' => 'Limit',
-    'description-limit' => 'Number of results to return'
+    'description-limit' => 'Number of results to return',
+
+    'title-collection' => 'Collection',
+    'description-collection' => 'Override pagination limits and return the entire collection [!USE WITH CARE!]',
 ];