+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The complete changelog for the Costs to Expect REST API, follows the format defined at https://keepachangelog.com/en/1.0.0/
 
+## [v1.19.1] - 2019-08-09
+### Added
+- We have added a `collection` parameter which is available to some collections. When the `collection` parameter is defined and `true`, it overrides the pagination limits, and we return the entire collection.
+
+### Changed
+- We have updated the pagination helper class to check to ensure that the `collection` parameter is allowable for the request.
+
 ## [v1.19.0] - 2019-08-08
 ### Added
 - We have added an `X-Sort` header to responses, displays the valid sort options, the applied sort options may differ to the requested sort options.